cosa/requirements.txt
numpy
pytorch
pygame